restrict access to edit and delete via controller

+++ app/controllers/articles_controller.rb
@@ -25,23 +25,38 @@ class ArticlesController < ApplicationController
 
   def show; end
 
-  def edit; end
+  def edit
+    unless @article.user == current_user
+      flash[:alert] = "You can only edit your own article"
+      redirect_to root_path
+    end
+  end
 
   def update
-    if @article.update(article_params)
-      flash[:success] = "Article has been updated"
-      redirect_to article_path(@article)
+    unless @article.user == current_user
+      flash[:alert] = "You can only edit your own article"
+      redirect_to root_path
     else
-      # We add flash.now to prevent this flash being held in the queue (given we render not redirect)
-      flash.now[:danger] = "Article has not been updated"
-      render :edit
+      if @article.update(article_params)
+        flash[:success] = "Article has been updated"
+        redirect_to article_path(@article)
+      else
+        # We add flash.now to prevent this flash being held in the queue (given we render not redirect)
+        flash.now[:danger] = "Article has not been updated"
+        render :edit
+      end
     end
   end
 
   def destroy
-    if @article.destroy
-      flash[:success] = "Article has been deleted"
-      redirect_to articles_path
+    unless @article.user == current_user
+      flash[:alert] = "You can only delete your own article"
+      redirect_to root_path
+    else
+      if @article.destroy
+        flash[:success] = "Article has been deleted"
+        redirect_to articles_path
+      end
     end
   end
